+++ AWSAPIGateway/AWSAPIGatewayClient.m
@@ -23,7 +23,7 @@ NSString *const AWSAPIGatewayErrorHTTPHeaderFieldsKey = @"HTTPHeaderFields";
 
 static NSString *const AWSAPIGatewayAPIKeyHeader = @"x-api-key";
 
-static NSString *const AWSAPIGatewaySDKVersion = @"2.4.2";
+static NSString *const AWSAPIGatewaySDKVersion = @"2.4.3";
 
 @interface AWSAPIGatewayClient()