+++ AWSLambda/AWSLambdaService.m
@@ -27,7 +27,7 @@
 #import "AWSLambdaResources.h"
 
 static NSString *const AWSInfoLambda = @"Lambda";
-static NSString *const AWSLambdaSDKVersion = @"2.4.2";
+static NSString *const AWSLambdaSDKVersion = @"2.4.3";
 
 @interface AWSLambdaResponseSerializer : AWSJSONResponseSerializer