+++ AWSMobileAnalytics/AWSMobileAnalyticsERS/AWSMobileAnalyticsERSService.m
@@ -28,7 +28,7 @@
 #import "AWSMobileAnalyticsERSResources.h"
 
 static NSString *const AWSInfoMobileAnalyticsERS = @"MobileAnalyticsERS";
-static NSString *const AWSMobileAnalyticsERSSDKVersion = @"2.4.2";
+static NSString *const AWSMobileAnalyticsERSSDKVersion = @"2.4.3";
 
 @interface AWSMobileAnalyticsERSResponseSerializer : AWSJSONResponseSerializer