+++ AWSSimpleDB/AWSSimpleDBService.m
@@ -27,7 +27,7 @@
 #import "AWSSimpleDBResources.h"
 
 static NSString *const AWSInfoSimpleDB = @"SimpleDB";
-static NSString *const AWSSimpleDBSDKVersion = @"2.4.2";
+static NSString *const AWSSimpleDBSDKVersion = @"2.4.3";
 
 @interface AWSSimpleDBResponseSerializer : AWSXMLResponseSerializer